Updated button colors and reorganized Products
