+++ client/components/BookHotel.css
@@ -4,66 +4,6 @@
   justify-content: space-between;
 }
 
-.product-card {
-  width: 300px;
-  height: 450px;
-  background-color: #fff;
-  border-radius: 8px;
-  box-shadow: 0px 8px 8px rgba(0, 0, 0, 0.2);
-  margin-bottom: 20px;
-  overflow: hidden;
-  transition: transform 0.3s ;
-  margin: 20px;
-  position: relative;
-}
-
-.book-now-button {
-  position: absolute;
-  bottom: 10px; /* Adjust the value as needed */
-  left: 50%; /* Adjust the value as needed */
-  transform: translateX(-50%); /* Add this line to center the button horizontally */
-  padding: 10px 0px; /* Add padding to the button */
-  background-color: #dbbaba;
-  color: #333;
-  border-radius: 8px;
-  text-decoration: none;
-
-}
-
-.book-now-button:hover {
-  background-color: #df9393;
-}
-
-.product-card:hover {
-  transform: translateY(-10px);
-}
-
-.product-image img {
-  width: 100%;
-  height: 150px;
-  object-fit: cover;
-}
-
-.product-info {
-  padding: 20px;
-}
-
-.product-title {
-  font-size: 1.2em;
-  margin-bottom: 22px;
-}
-
-.product-price { 
-    color: black;
-    margin-bottom: 10px;
-  
-}
-
-.product-description {
-  font-size: 0.9em;
-  margin-bottom: 0px;
-}
-
 .hotel-container {
 background-color: #dbbaba;
 padding: 30px;
